Add status helper text and update toast message for status change
